trap: give disasm some nop's, so it shows the label
the disasm doesn't show labels on invalid opcode, but
that's convenient when debugging.

one can argue it's not really efficient here, because two cycles are wastet.
imho that's not true, since modern x86 cpu's eliminate such instructions
sequences.

we could benchmark/test that, but well, it's just on the first hit, so suck it.