exception: mark each IRLabel with exception information
in the current design, we can't easily drag JVM PC information nicely with every
instruction. Fortunately, it should be enough to have information about it on
basic block level, i.e. every block knows about itself:
	(1) which exception can occur in this block
	(2) is this block a handler

probably this could be a problem in the future, e.g. precise information which
statement caused an exception (but, it should be enough for stacktraces though)

Also note the usage of `IntervalMap'. I don't know why I didn't found it in the
first place.