HVar: remove SpillRReg, and rename it to HVarX86
with SpillRReg it was too genral and it makes the codegen quite verbose. so
remove it for x86.

the only point in codegen where it could be relevant is for GCPoints, but we
have there type information anyways. I compared the GCPoints in one testcase and
it's still the same.