hoopl: introducing register coalescing pass
this pass tries to merge virtual registers and therefore eliminate move
instructions, which were generated by MkGraph.

For some cases, we have to be carefully, e.g. when a new value is assigned to an
argument register. Thus, some hard-to-handle-situations are avoided, by not
applying the optimization under certain conditions.