classpath debugging: add printInteger
this one is tricky: if you use something like
	jmate.lang.System.out.printf("sup: %d\n", 1337);
`javac' generates code which includes `new Integer(1337)', in order to pass
varargs to printf. However, this implies to use java.lang.Integer by GNU
Classpath, which in turn has several side-effects, resulting in a nullpointer
exception somewhere.

The solution is to avoid java.lang.Integer, and use a hackish variant in jmate.*