liveness: refactor LiveAnnotation away from IR constructors
I came to the point to describe the intermediate language in my thesis, and
thought it's ugly to have the liveness annotation encoded in every constructor.

this is a different approach, so MateIR definition is clean again. it is a bit
more complicated on the typelevel (see LivenessPass.hs and Linear.hs), but I
think it's worth it.